Add a first submission skeleton.
The notebook initially provided is a bit bloated, does not handle predictions on test data and does not create a submission file.
This one is cleaner (it uses helpers.py and plots.py) and it creates a valid submission file from predictions.
Note that the model performance has been overlooked for now.